rm import from servicelib
