+++ packages/models-library/src/models_library/api_schemas_payments/__init__.py
@@ -1,6 +1,7 @@
 from typing import Final
 
 from pydantic import parse_obj_as
-from servicelib.rabbitmq import RPCNamespace
+
+from ..rabbitmq_basic_types import RPCNamespace
 
 PAYMENTS_RPC_NAMESPACE: Final[RPCNamespace] = parse_obj_as(RPCNamespace, "payments")
