+++ packages/service-library/src/servicelib/rabbitmq/_constants.py
@@ -1,4 +1,3 @@
 from typing import Final
 
 BIND_TO_ALL_TOPICS: Final[str] = "#"
-REGEX_RABBIT_QUEUE_ALLOWED_SYMBOLS: Final[str] = r"^[\w\-\.]*$"
