+++ packages/service-library/src/servicelib/rabbitmq/_models.py
@@ -2,9 +2,11 @@ import re
 from collections.abc import Awaitable, Callable
 from typing import Any, Protocol
 
+from models_library.rabbitmq_basic_types import (
+    REGEX_RABBIT_QUEUE_ALLOWED_SYMBOLS,
+    RPCNamespace,
+)
 from pydantic import ConstrainedStr, parse_obj_as
-
-from ._constants import REGEX_RABBIT_QUEUE_ALLOWED_SYMBOLS
 
 MessageHandler = Callable[[Any], Awaitable[bool]]
 
@@ -23,21 +25,6 @@ class RPCMethodName(ConstrainedStr):
     regex: re.Pattern[str] | None = re.compile(REGEX_RABBIT_QUEUE_ALLOWED_SYMBOLS)
 
 
-class RPCNamespace(ConstrainedStr):
-    min_length: int = 1
-    max_length: int = 252
-    regex: re.Pattern[str] | None = re.compile(REGEX_RABBIT_QUEUE_ALLOWED_SYMBOLS)
-
-    @classmethod
-    def from_entries(cls, entries: dict[str, str]) -> "RPCNamespace":
-        """
-        Given a list of entries creates a namespace to be used in declaring the rabbitmq queue.
-        Keeping this to a predefined length
-        """
-        composed_string = "-".join(f"{k}_{v}" for k, v in sorted(entries.items()))
-        return parse_obj_as(cls, composed_string)
-
-
 class RPCNamespacedMethodName(ConstrainedStr):
     min_length: int = 1
     max_length: int = 255
@@ -49,3 +36,6 @@ class RPCNamespacedMethodName(ConstrainedStr):
     ) -> "RPCNamespacedMethodName":
         namespaced_method_name = f"{namespace}.{method_name}"
         return parse_obj_as(cls, namespaced_method_name)
+
+
+__all__: tuple[str, ...] = ("RPCNamespace",)
